Move ResetVector etc. templates into new Helper.h header
Also, add a couple of other useful templates:
 DeleteContainer
 DeleteContainerOfContainers
which can be used to delete vectors and other STL containers holding
pointers to objects that they own.
