+++ SBSCalorimeter.h
@@ -19,6 +19,7 @@
 #include "TRotation.h"
 #include "TVector3.h"
 #include "THaDetMap.h"
+#include "Helper.h"
 
 struct SBSBlockSet {
   Double_t e;
@@ -150,40 +151,6 @@ protected:
 
   Double_t GetVVal(std::vector<Double_t> &v, UInt_t i = 0 );
   Int_t GetVVal(std::vector<Int_t> &v, UInt_t i = 0 );
-
-private:
-  // Simple and quick routine to init and clear most vectors
-  // (of integers, doubles, doubles, etc...)
-  // Reset/Init 1D vector
-  template<class T>
-  void InitVector(std::vector<T> &vec, T val = 0, size_t n = 0) {
-    vec.resize(n);
-    ResetVector(vec,val);
-  }
-
-  template<class T>
-  void ResetVector(std::vector<T> &vec, T val = 0, size_t n = 0) {
-    if(n > 0) {
-      vec.clear();
-      vec.resize(n);
-    }
-    for(size_t i = 0; i < vec.size(); i++) {
-      vec[i] = val;
-    }
-  }
-
-  // Reset 2D vector
-  template<class T>
-  void ResetVector(std::vector<std::vector<T> > &vec, T val = 0,
-      size_t nr = 0, size_t nc = 0) {
-    if(nr > 0) {
-      vec.clear();
-      vec.resize(nr);
-    }
-    for(size_t i = 0; i < vec.size(); i++) {
-      ResetVector(vec[i],val,nc);
-    }
-  }
 
   ClassDef(SBSCalorimeter,0)     //Generic shower detector class
 };
